+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.0
+
+* Breaking change: The type of rebuild is changed from Stream<LoadingMoreBase<T>>  to Stream<Iterable<T>> in LoadingMoreBase<T>.
+* Add a new method 'wrapData', you can modify the list here.
+
 ## 2.0.1
 
 * Fix error when add non-null T